Update env/settings samples and README to use Azurite by default
