+++ function_app/sample_local.settings.json
@@ -4,10 +4,6 @@
         "FUNCTIONS_WORKER_RUNTIME": "python",
         "AzureWebJobsFeatureFlags": "EnableWorkerIndexing",
         "AzureWebJobsStorage": "UseDevelopmentStorage=true",
-        "AzureWebJobsStorage__accountName": "Copy `storageAccountName` from `.azure/<env>/.env` - E.g: 'llm-proc-storage-xxxxx'",
-        "AzureWebJobsStorage__blobServiceUri": "Copy `storageAccountBlobUri` from `.azure/<env>/.env` - E.g: 'https://<storage-account-name>.blob.core.windows.net/'",
-        "AzureWebJobsStorage__queueServiceUri": "Copy `storageAccountQueueUri` from `.azure/<env>/.env` - E.g: 'https://<storage-account-name>.queue.core.windows.net/'",
-        "AzureWebJobsStorage__tableServiceUri": "Copy `storageAccountTableUri` from `.azure/<env>/.env` - E.g: 'https://<storage-account-name>.table.core.windows.net/'",
         "CosmosDbConnectionSetting__accountEndpoint": "Copy `cosmosDbAccountEndpoint` from `.azure/<env>/.env` - E.g: 'https://<cosmosdb_endpoint>.documents.azure.com:443/'",
         "COSMOSDB_ACCOUNT": "Copy `cosmosDbAccountName` from `.azure/<env>/.env` - E.g: 'llm-proc-cosmosdb-xxxxx'",
         "COSMOSDB_DATABASE_NAME": "Copy `cosmosDbDatabaseName` from `.azure/<env>/.env` - E.g: 'default'",
